Fix the high CPU issue
It's because we updaing the widget every timeout
and the QPainter.drawText() is very time consuming.
So when timeout, we check if the spans is changed,
only when changed then we actually do the redrawing.

The other files are unsuccessful attempt to rewrite
this project.
Well, with now the issue gone, I find myself losing
the enthusiasm to keep on going, especially in this
special period when I need focus more on the preparing
for the post-graduate things.
And i'm kinda getting a fever i feel, it's not comfortable.