Put line break at end of session instead of head
Vim will automatically put an end-of-line at every
file it save.
This behavior will break the proper functioning of
my program because it expect to find no eol at the
end of file.
With no easy way to disable this and with no intent
to not use vim, i choose to fiddle a little in the
code to fix this.
Voola!